+++ quilt-sync/ui/src/components/buttons/pull.rs
@@ -11,9 +11,12 @@ pub fn Pull(
     on_click: impl Fn(leptos::ev::MouseEvent) + 'static,
     #[prop(optional)] small: bool,
     #[prop(optional, into)] busy: MaybeProp<bool>,
+    #[prop(optional, into)] disabled: MaybeProp<bool>,
 ) -> impl IntoView {
+    let is_disabled =
+        Signal::derive(move || busy.get().unwrap_or(false) || disabled.get().unwrap_or(false));
     view! {
-        <IconButton icon=KIND.icon() on_click=UnsyncCallback::new(on_click) small=small primary=true disabled=busy>
+        <IconButton icon=KIND.icon() on_click=UnsyncCallback::new(on_click) small=small primary=true disabled=is_disabled>
             {move || if busy.get().unwrap_or(false) { "Pulling\u{2026}" } else { KIND.label() }}
         </IconButton>
     }